# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DICOM Gateway API
  description: |
    REST API for DICOM Gateway - A lightweight, HIPAA-compliant Linux DICOM Gateway
    for receiving and forwarding DICOM studies with byte-preserving integrity.
    
    ## Features
    - C-STORE SCP (receiver) and SCU (forwarder) with byte-preserving transmission
    - PostgreSQL-backed job queue with LISTEN/NOTIFY
    - Role-based access control (RBAC)
    - Comprehensive audit logging
    - Prometheus metrics export
    - Configuration management via YAML
  version: 0.1.0
  contact:
    name: DICOM Gateway Support
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://gateway.example.com/api/v1
    description: Production server
...
paths:
...
  /studies:
    get:
      summary: List studies
      description: List DICOM studies with pagination and filtering
      tags:
        - Studies
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Skip'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          schema:
            type: string
          description: Filter by status
        - name: patient_id
          in: query
          schema:
            type: string
          description: Filter by patient ID
        - name: study_date
          in: query
          schema:
            type: string
            format: date
          description: Filter by study date
      responses:
        '200':
          description: List of studies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudySummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /studies/{study_id}:
    get:
      summary: Get study by ID
      description: Get detailed study information by ID
      tags:
        - Studies
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StudyId'
      responses:
        '200':
          description: Study details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudyDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /studies/uid/{study_instance_uid}:
    get:
      summary: Get study by UID
      description: Get detailed study information by Study Instance UID
      tags:
        - Studies
      security:
        - BearerAuth: []
      parameters:
        - name: study_instance_uid
          in: path
          required: true
          schema:
            type: string
          description: Study Instance UID
      responses:
        '200':
          description: Study details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudyDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /studies/{study_id}/forward:
    post:
      summary: Forward study
      description: Create forward jobs for a study to specified destinations
      tags:
        - Studies
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StudyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForwardRequest'
      responses:
        '200':
          description: Forward jobs created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  jobs_created:
                    type: integer
                  job_ids:
                    type: array
                    items:
                      type: string
                      format: uuid
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Skip:
      name: skip
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Number of records to skip
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
      description: Maximum number of records to return
    StudyId:
      name: study_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Study ID
...
  schemas:
...
    StudySummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        study_instance_uid:
          type: string
        patient_id:
          type: string
          nullable: true
        patient_name:
          type: string
          nullable: true
        study_date:
          type: string
          nullable: true
        accession_number:
          type: string
          nullable: true
        modality:
          type: string
          nullable: true
        status:
          type: string
        file_count:
          type: integer
        total_size_bytes:
          type: integer
        created_at:
          type: string
          format: date-time
    StudyDetail:
      allOf:
        - $ref: '#/components/schemas/StudySummary'
        - type: object
          properties:
            study_time:
              type: string
              nullable: true
            study_description:
              type: string
              nullable: true
            referring_physician_name:
              type: string
              nullable: true
            institution_name:
              type: string
              nullable: true
            forwarded_at:
              type: string
              format: date-time
              nullable: true
    ForwardRequest:
      type: object
      required:
        - destination_ids
      properties:
        destination_ids:
          type: array
          items:
            type: string
            format: uuid
        priority:
          type: integer
          default: 0
...
    Error:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'